+++ code/src/cljs/sixsq/nuvla/ui/profile/spec.cljs
@@ -1,6 +1,7 @@
 (ns sixsq.nuvla.ui.profile.spec
   (:require
-    [clojure.spec.alpha :as s]))
+    [clojure.spec.alpha :as s]
+    [sixsq.nuvla.ui.utils.spec :as us]))
 
 
 (s/def ::user any?)
@@ -31,6 +32,18 @@
 
 (s/def ::vendor any?)
 
+(s/def ::group any?)
+
+(s/def ::group-name us/nonblank-string)
+
+(s/def ::group-description us/nonblank-string)
+
+(s/def ::group-form (s/keys :req [::group-name
+                                  ::group-description]))
+
+
+(s/def ::active-tab-index number?)
+
 
 (s/def ::db (s/keys :req [::user
                           ::customer
@@ -44,7 +57,9 @@
                           ::error-message
                           ::loading
                           ::setup-intent
-                          ::vendor]))
+                          ::vendor
+                          ::group
+                          ::active-tab-index]))
 
 
 (def defaults {::user              nil
@@ -59,4 +74,6 @@
                ::open-modal        nil
                ::error-message     nil
                ::loading           #{}
-               ::vendor            nil})
+               ::vendor            nil
+               ::group             nil
+               ::active-tab-index  0})