+++ code/src/cljs/sixsq/nuvla/ui/session/events.cljs
@@ -10,7 +10,6 @@
     [sixsq.nuvla.ui.history.events :as history-events]
     [sixsq.nuvla.ui.intercom.events :as intercom-events]
     [sixsq.nuvla.ui.main.spec :as main-spec]
-    [sixsq.nuvla.ui.profile.events :as profile-events]
     [sixsq.nuvla.ui.session.effects :as fx]
     [sixsq.nuvla.ui.session.spec :as spec]
     [sixsq.nuvla.ui.utils.response :as response]))
@@ -25,7 +24,9 @@
                               (when session
                                 #_(dispatch [:sixsq.nuvla.ui.main.events/check-bootstrap-message])
                                 (dispatch [:sixsq.nuvla.ui.main.events/notifications-polling])
-                                (dispatch [::profile-events/search-existing-customer])))]}))
+                                (dispatch [:sixsq.nuvla.ui.profile.events/search-existing-customer])
+                                (dispatch [::search-groups])
+                                (dispatch [::get-peers])))]}))
 
 
 (reg-event-fx
@@ -36,12 +37,14 @@
     (let [no-session-protected-page? (and (nil? session-arg)
                                           (->> nav-path first (get pages) :protected?))]
       (cond-> {:db (assoc db ::spec/session session-arg
-                             ::spec/session-loading? false)}
-              no-session-protected-page? (assoc :dispatch [::history-events/navigate "sign-in"])
+                             ::spec/session-loading? false
+                             :dispatch-n [])}
               session-arg (assoc ::fx/automatic-logout-at-session-expiry [session-arg])
+              no-session-protected-page? (update :dispatch-n conj [::history-events/navigate "sign-in"])
               ;; force refresh templates collection cache when not the same user (different session)
-              (not= session session-arg) (assoc :dispatch-n [[::cimi-events/get-cloud-entry-point]
-                                                             [:sixsq.nuvla.ui.main.events/force-refresh-content]])))))
+              (not= session session-arg) (update :dispatch-n conj
+                                                 [::cimi-events/get-cloud-entry-point]
+                                                 [:sixsq.nuvla.ui.main.events/force-refresh-content])))))
 
 
 (reg-event-fx
@@ -177,3 +180,39 @@
     (let [claim (if (= (:identifier session) claim) (:user session) claim)]
       {::cimi-api-fx/operation [(:id session) "switch-group" #(dispatch [::initialize])
                                 {:claim claim}]})))
+
+
+(reg-event-db
+  ::set-peers
+  (fn [db [_ response]]
+    (if (instance? js/Error response)
+      (let [{:keys [status message]} (response/parse-ex-info response)]
+        (js/console.error "Get peers failed (" status "): " message)
+        db)
+      (assoc db ::spec/peers
+                (->> response
+                     (map (fn [[k v]] [(str (namespace k) "/" (name k)) v]))
+                     (sort-by (juxt second first))
+                     (into {}))))))
+
+
+(reg-event-fx
+  ::get-peers
+  (fn [{{:keys [::spec/session]} :db}]
+    {::cimi-api-fx/operation [(:id session) "get-peers" #(dispatch [::set-peers %])]}))
+
+
+(reg-event-db
+  ::set-groups
+  (fn [db [_ {:keys [resources]}]]
+    (when resources
+      (assoc db ::spec/groups resources))))
+
+
+(reg-event-fx
+  ::search-groups
+  (fn [_ _]
+    {::cimi-api-fx/search [:group {:select  "id, name, acl, users, description"
+                                   :last    10000
+                                   :orderby "name:asc,id:asc"}
+                           #(dispatch [::set-groups %])]}))